move DEMLayer code to AbstractDEMLayer
